Port xmonad keybindings to hyprland config
The hyprland config is an attempt to port the existing XMonad
configuration to hyprland. Most of it can be handled with the
available configuration options, however keeping the window
behaviour required a small bash script taken from:

https://codeberg.org/JustineSmithies/hyprland-dotfiles/src/branch/master/.config/eww/scripts/dispatch.sh

The ability to treat fullscreen as a separate layout also required a
bash script. It determines the existing layout, and rotates to the next
one. If it is rotating to fullscreen then it dispatches the fullscreen
command instead.